pypi(client): move client into own submodule
