+++ pypi/clients/__init__.py
@@ -1,10 +1,11 @@
-"""This module defines code for interacting with the PyPi APIs
+"""This module contains clients for interacting with the PyPi APIs
 
 Attributes:
     PypiIndexClient: Client for interacting with the Index API
     PypiJsonClient: Client for interacting with the JSON API
 """
 
-from .clients import PypiIndexClient, PypiJsonClient
+from .index import PypiIndexClient
+from .json import PypiJsonClient
 
 __all__ = ["PypiIndexClient", "PypiJsonClient"]
